Use environment variables for database details
Jira ticket: CAMS-207
